requirements.txt
oracledb
simple-salesforce
python-dotenv
pandas
rapidfuzz
tqdm
pyyaml